+++ CHANGELOG.md
@@ -13,7 +13,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Fixed an issue on iOS where fetching the artwork for the NowPlayingInfo could cause a crash due to unsafe thread usage.
+
+## [11.4.1] - 26-07-22
+
+### Changed
+
+- Added RCTBridgeModule extension on iOS, with helper methods to resolve the view and native player APIs.
+
+### Fixed
+
 - Fixed an issue on iOS where sideloaded texttracks were not stored for cached playout.
+
+### Added
+
+- Added `contentprotectionerror` player event on Web, Android, and iOS platforms.
 
 ## [11.4.0] - 26-07-06